requirements.txt
streamlit
google-generativeai
requests
duckduckgo-search
beautifulsoup4
psutil
wikipedia
yfinance
python-dotenv
